+++ src/Discovery/BindingTypeDescriptorStore.php
@@ -9,9 +9,8 @@
  * file that was distributed with this source code.
  */
 
-namespace Puli\RepositoryManager\Discovery\Store;
+namespace Puli\RepositoryManager\Discovery;
 
-use Puli\RepositoryManager\Discovery\BindingTypeDescriptor;
 use Puli\RepositoryManager\Package\Package;
 use Puli\RepositoryManager\Util\CompositeKeyStore;
 
@@ -19,7 +18,7 @@ use Puli\RepositoryManager\Util\CompositeKeyStore;
  * @since  1.0
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-class BindingTypeStore
+class BindingTypeDescriptorStore
 {
     /**
      * @var CompositeKeyStore